Test parsing json file

+++ onvm_web/cors_server.py
@@ -6,7 +6,6 @@ import json
 import subprocess
 import os
 import signal
-import shlex
 
 global is_running
 is_running = -1
@@ -63,6 +62,9 @@ class CORSRequestHandler(SimpleHTTPRequestHandler):
                 self.wfile.write(str.encode(response))
                 # setup return result
                 return -1
+
+            self.parst_request_file(request_body_json)
+
             # start the process and change the state
             log_file = open('./test.txt', 'w+')
             os.chdir('../examples/')
@@ -88,7 +90,6 @@ class CORSRequestHandler(SimpleHTTPRequestHandler):
 
     def stop_nf_chain(self):
         global is_running
-        message = []
         try:
             # open the log file to read the process name of the nfs
             with open('./test.txt', 'r') as log_file:
@@ -97,19 +98,18 @@ class CORSRequestHandler(SimpleHTTPRequestHandler):
                     log_info = log.split(" ")
                     if(log_info[0] == "Starting"):
                         command = "ps -ef | grep sudo | grep " + log_info[1] + " | grep -v 'grep' | awk '{print $2}'"
-                        # command = ["ps", "-ef", "|", "grep", "sudo", "|", "grep", log_info[1], "|", "grep", "-v", "'grep'", "|", "awk", "'{print $2}'"]
                         pids = os.popen(command)
                         pid_processes = pids.read()
                         if pid_processes != "":
                             pid_processes = pid_processes.split("\n")
-                            message.append(pid_processes)
                             for i in pid_processes:
-                                os.kill(int(i), signal.SIGKILL)
+                                if i != "":
+                                    os.kill(int(i), signal.SIGKILL)
                     log = log_file.readline()
             # reset is_running
             is_running = -1
             self.send_response(200)
-            response = json.dumps({'status': '200', 'message': 'stop nfs successed', "pids": message})
+            response = json.dumps({'status': '200', 'message': 'stop nfs successed'})
             result = 0
         except OSError:
             self.send_response(500)
@@ -131,7 +131,9 @@ class CORSRequestHandler(SimpleHTTPRequestHandler):
             response = json.dumps({'status': '500', 'message': 'missing data'})
             self.wfile.write(str.encode(response))
             return None
-        print(type(data))
+        with open('./log.txt', 'w+') as log_file:
+            log_file.write(type(data))
+            log_file.write("\n" + data)
 
 if __name__ == '__main__':
     test(CORSRequestHandler, HTTPServer, port=8000)
